requirements.txt
pillow>=8,<9
